% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{AffineTransform}
\alias{AffineTransform}
\title{AffineTransform Object}
\usage{
AffineTransform(unit = NULL, scaleX = NULL, shearX = NULL,
  scaleY = NULL, translateY = NULL, translateX = NULL, shearY = NULL)
}
\arguments{
\item{unit}{The units for translate elements}

\item{scaleX}{The X coordinate scaling element}

\item{shearX}{The X coordinate shearing element}

\item{scaleY}{The Y coordinate scaling element}

\item{translateY}{The Y coordinate translation element}

\item{translateX}{The X coordinate translation element}

\item{shearY}{The Y coordinate shearing element}
}
\value{
AffineTransform object
}
\description{
AffineTransform Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
AffineTransform uses a 3x3 matrix with an implied last row of [ 0 0 1 ]to transform source coordinates (x,y) into destination coordinates (x', y')according to:      x'  x  =   shear_y  scale_y  translate_y        1  [ 1 ]After transformation,     x' = scale_x * x + shear_x * y + translate_x;     y' = scale_y * y + shear_y * x + translate_y;This message is therefore composed of these six matrix elements.
}
