% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{AutoText}
\alias{AutoText}
\title{AutoText Object}
\usage{
AutoText(style = NULL, type = NULL, content = NULL)
}
\arguments{
\item{style}{The styling applied to this auto text}

\item{type}{The type of this auto text}

\item{content}{The rendered content of this auto text, if available}
}
\value{
AutoText object
}
\description{
AutoText Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TextElement kind that represents auto text.
}
