% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{BatchUpdatePresentationRequest}
\alias{BatchUpdatePresentationRequest}
\title{BatchUpdatePresentationRequest Object}
\usage{
BatchUpdatePresentationRequest(requests = NULL)
}
\arguments{
\item{requests}{A list of updates to apply to the presentation}
}
\value{
BatchUpdatePresentationRequest object
}
\description{
BatchUpdatePresentationRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Request message for PresentationsService.BatchUpdatePresentation.
}
\seealso{
Other BatchUpdatePresentationRequest functions: \code{\link{presentations.batchUpdate}}
}
