% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{BatchUpdatePresentationResponse}
\alias{BatchUpdatePresentationResponse}
\title{BatchUpdatePresentationResponse Object}
\usage{
BatchUpdatePresentationResponse(presentationId = NULL, replies = NULL)
}
\arguments{
\item{presentationId}{The presentation the updates were applied to}

\item{replies}{The reply of the updates}
}
\value{
BatchUpdatePresentationResponse object
}
\description{
BatchUpdatePresentationResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Response message from a batch update.
}
