% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Bullet}
\alias{Bullet}
\title{Bullet Object}
\usage{
Bullet(glyph = NULL, nestingLevel = NULL, bulletStyle = NULL,
  listId = NULL)
}
\arguments{
\item{glyph}{The rendered bullet glyph for this paragraph}

\item{nestingLevel}{The nesting level of this paragraph in the list}

\item{bulletStyle}{The paragraph specific text style applied to this bullet}

\item{listId}{The ID of the list this paragraph belongs to}
}
\value{
Bullet object
}
\description{
Bullet Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Describes the bullet of a paragraph.
}
