% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ColorScheme}
\alias{ColorScheme}
\title{ColorScheme Object}
\usage{
ColorScheme(colors = NULL)
}
\arguments{
\item{colors}{The ThemeColorType and corresponding concrete color pairs}
}
\value{
ColorScheme object
}
\description{
ColorScheme Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The palette of predefined colors for a page.
}
