% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ColorStop}
\alias{ColorStop}
\title{ColorStop Object}
\usage{
ColorStop(alpha = NULL, position = NULL, color = NULL)
}
\arguments{
\item{alpha}{The alpha value of this color in the gradient band}

\item{position}{The relative position of the color stop in the gradient band measured}

\item{color}{The color of the gradient stop}
}
\value{
ColorStop object
}
\description{
ColorStop Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A color and position in a gradient band.
}
