% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CreateImageRequest}
\alias{CreateImageRequest}
\title{CreateImageRequest Object}
\usage{
CreateImageRequest(objectId = NULL, elementProperties = NULL, url = NULL)
}
\arguments{
\item{objectId}{A user-supplied object ID}

\item{elementProperties}{The element properties for the image}

\item{url}{The image URL}
}
\value{
CreateImageRequest object
}
\description{
CreateImageRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creates an image.
}
