% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CreateImageResponse}
\alias{CreateImageResponse}
\title{CreateImageResponse Object}
\usage{
CreateImageResponse(objectId = NULL)
}
\arguments{
\item{objectId}{The object ID of the created image}
}
\value{
CreateImageResponse object
}
\description{
CreateImageResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of creating an image.
}
