% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CreateLineRequest}
\alias{CreateLineRequest}
\title{CreateLineRequest Object}
\usage{
CreateLineRequest(objectId = NULL, elementProperties = NULL,
  lineCategory = NULL)
}
\arguments{
\item{objectId}{A user-supplied object ID}

\item{elementProperties}{The element properties for the line}

\item{lineCategory}{The category of line to be created}
}
\value{
CreateLineRequest object
}
\description{
CreateLineRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creates a line.
}
