% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CreateLineResponse}
\alias{CreateLineResponse}
\title{CreateLineResponse Object}
\usage{
CreateLineResponse(objectId = NULL)
}
\arguments{
\item{objectId}{The object ID of the created line}
}
\value{
CreateLineResponse object
}
\description{
CreateLineResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of creating a line.
}
