% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CreateParagraphBulletsRequest}
\alias{CreateParagraphBulletsRequest}
\title{CreateParagraphBulletsRequest Object}
\usage{
CreateParagraphBulletsRequest(objectId = NULL, textRange = NULL,
  bulletPreset = NULL, cellLocation = NULL)
}
\arguments{
\item{objectId}{The object ID of the shape or table containing the text to add bullets to}

\item{textRange}{The range of text to apply the bullet presets to, based on TextElement indexes}

\item{bulletPreset}{The kinds of bullet glyphs to be used}

\item{cellLocation}{The optional table cell location if the text to be modified is in a table}
}
\value{
CreateParagraphBulletsRequest object
}
\description{
CreateParagraphBulletsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creates bullets for all of the paragraphs that overlap with the giventext index range.The nesting level of each paragraph will be determined by counting leadingtabs in front of each paragraph. To avoid excess space between the bullet andthe corresponding paragraph, these leading tabs are removed by this request.This may change the indices of parts of the text.If the paragraph immediately before paragraphs being updated is in a listwith a matching preset, the paragraphs being updated are added to thatpreceding list.
}
