% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CreateShapeRequest}
\alias{CreateShapeRequest}
\title{CreateShapeRequest Object}
\usage{
CreateShapeRequest(objectId = NULL, shapeType = NULL,
  elementProperties = NULL)
}
\arguments{
\item{objectId}{A user-supplied object ID}

\item{shapeType}{The shape type}

\item{elementProperties}{The element properties for the shape}
}
\value{
CreateShapeRequest object
}
\description{
CreateShapeRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creates a new shape.
}
