% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CreateShapeResponse}
\alias{CreateShapeResponse}
\title{CreateShapeResponse Object}
\usage{
CreateShapeResponse(objectId = NULL)
}
\arguments{
\item{objectId}{The object ID of the created shape}
}
\value{
CreateShapeResponse object
}
\description{
CreateShapeResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of creating a shape.
}
