% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CreateSheetsChartResponse}
\alias{CreateSheetsChartResponse}
\title{CreateSheetsChartResponse Object}
\usage{
CreateSheetsChartResponse(objectId = NULL)
}
\arguments{
\item{objectId}{The object ID of the created chart}
}
\value{
CreateSheetsChartResponse object
}
\description{
CreateSheetsChartResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of creating an embedded Google Sheets chart.
}
