% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CreateSlideRequest}
\alias{CreateSlideRequest}
\title{CreateSlideRequest Object}
\usage{
CreateSlideRequest(slideLayoutReference = NULL, objectId = NULL,
  insertionIndex = NULL, placeholderIdMappings = NULL)
}
\arguments{
\item{slideLayoutReference}{Layout reference of the slide to be inserted, based on the *current}

\item{objectId}{A user-supplied object ID}

\item{insertionIndex}{The optional zero-based index indicating where to insert the slides}

\item{placeholderIdMappings}{An optional list of object ID mappings from the placeholder(s) on the layout to the placeholder(s)}
}
\value{
CreateSlideRequest object
}
\description{
CreateSlideRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creates a new slide.
}
