% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CreateTableRequest}
\alias{CreateTableRequest}
\title{CreateTableRequest Object}
\usage{
CreateTableRequest(objectId = NULL, columns = NULL,
  elementProperties = NULL, rows = NULL)
}
\arguments{
\item{objectId}{A user-supplied object ID}

\item{columns}{Number of columns in the table}

\item{elementProperties}{The element properties for the table}

\item{rows}{Number of rows in the table}
}
\value{
CreateTableRequest object
}
\description{
CreateTableRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creates a new table.
}
