% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CreateTableResponse}
\alias{CreateTableResponse}
\title{CreateTableResponse Object}
\usage{
CreateTableResponse(objectId = NULL)
}
\arguments{
\item{objectId}{The object ID of the created table}
}
\value{
CreateTableResponse object
}
\description{
CreateTableResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of creating a table.
}
