% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CreateVideoRequest}
\alias{CreateVideoRequest}
\title{CreateVideoRequest Object}
\usage{
CreateVideoRequest(objectId = NULL, source = NULL,
  elementProperties = NULL, id = NULL)
}
\arguments{
\item{objectId}{A user-supplied object ID}

\item{source}{The video source}

\item{elementProperties}{The element properties for the video}

\item{id}{The video source's unique identifier for this video}
}
\value{
CreateVideoRequest object
}
\description{
CreateVideoRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Creates a video.
}
