% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CreateVideoResponse}
\alias{CreateVideoResponse}
\title{CreateVideoResponse Object}
\usage{
CreateVideoResponse(objectId = NULL)
}
\arguments{
\item{objectId}{The object ID of the created video}
}
\value{
CreateVideoResponse object
}
\description{
CreateVideoResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of creating a video.
}
