% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{CropProperties}
\alias{CropProperties}
\title{CropProperties Object}
\usage{
CropProperties(bottomOffset = NULL, angle = NULL, topOffset = NULL,
  leftOffset = NULL, rightOffset = NULL)
}
\arguments{
\item{bottomOffset}{The offset specifies the bottom edge of the crop rectangle that is located}

\item{angle}{The rotation angle of the crop window around its center, in radians}

\item{topOffset}{The offset specifies the top edge of the crop rectangle that is located}

\item{leftOffset}{The offset specifies the left edge of the crop rectangle that is located to}

\item{rightOffset}{The offset specifies the right edge of the crop rectangle that is located}
}
\value{
CropProperties object
}
\description{
CropProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The crop properties of an object enclosed in a container. For example, anImage.The crop properties is represented by the offsets of four edges which definea crop rectangle. The offsets are measured in percentage from thecorresponding edges of the object's original bounding rectangle towardsinside, relative to the object's original dimensions.- If the offset is in the interval (0, 1), the corresponding edge of croprectangle is positioned inside of the object's original bounding rectangle.- If the offset is negative or greater than 1, the corresponding edge of croprectangle is positioned outside of the object's original bounding rectangle.- If the left edge of the crop rectangle is on the right side of its rightedge, the object will be flipped horizontally.- If the top edge of the crop rectangle is below its bottom edge, the objectwill be flipped vertically.- If all offsets and rotation angle is 0, the object is not cropped.After cropping, the content in the crop rectangle will be stretched to fitits container.
}
