% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{DeleteObjectRequest}
\alias{DeleteObjectRequest}
\title{DeleteObjectRequest Object}
\usage{
DeleteObjectRequest(objectId = NULL)
}
\arguments{
\item{objectId}{The object ID of the page or page element to delete}
}
\value{
DeleteObjectRequest object
}
\description{
DeleteObjectRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes an object, either pages orpage elements, from thepresentation.
}
