% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{DeleteParagraphBulletsRequest}
\alias{DeleteParagraphBulletsRequest}
\title{DeleteParagraphBulletsRequest Object}
\usage{
DeleteParagraphBulletsRequest(objectId = NULL, textRange = NULL,
  cellLocation = NULL)
}
\arguments{
\item{objectId}{The object ID of the shape or table containing the text to delete bullets}

\item{textRange}{The range of text to delete bullets from, based on TextElement indexes}

\item{cellLocation}{The optional table cell location if the text to be modified is in a table}
}
\value{
DeleteParagraphBulletsRequest object
}
\description{
DeleteParagraphBulletsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes bullets from all of the paragraphs that overlap with the given textindex range.The nesting level of each paragraph will be visually preserved by addingindent to the start of the corresponding paragraph.
}
