% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{DeleteTableRowRequest}
\alias{DeleteTableRowRequest}
\title{DeleteTableRowRequest Object}
\usage{
DeleteTableRowRequest(cellLocation = NULL, tableObjectId = NULL)
}
\arguments{
\item{cellLocation}{The reference table cell location from which a row will be deleted}

\item{tableObjectId}{The table to delete rows from}
}
\value{
DeleteTableRowRequest object
}
\description{
DeleteTableRowRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes a row from a table.
}
