% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{DeleteTextRequest}
\alias{DeleteTextRequest}
\title{DeleteTextRequest Object}
\usage{
DeleteTextRequest(cellLocation = NULL, objectId = NULL, textRange = NULL)
}
\arguments{
\item{cellLocation}{The optional table cell location if the text is to be deleted from a table}

\item{objectId}{The object ID of the shape or table from which the text will be deleted}

\item{textRange}{The range of text to delete, based on TextElement indexes}
}
\value{
DeleteTextRequest object
}
\description{
DeleteTextRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Deletes text from a shape or a table cell.
}
