% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Dimension}
\alias{Dimension}
\title{Dimension Object}
\usage{
Dimension(magnitude = NULL, unit = NULL)
}
\arguments{
\item{magnitude}{The magnitude}

\item{unit}{The units for magnitude}
}
\value{
Dimension object
}
\description{
Dimension Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A magnitude in a single direction in the specified units.
}
