% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{DuplicateObjectRequest}
\alias{DuplicateObjectRequest}
\title{DuplicateObjectRequest Object}
\usage{
DuplicateObjectRequest(DuplicateObjectRequest.objectIds = NULL,
  objectIds = NULL, objectId = NULL)
}
\arguments{
\item{DuplicateObjectRequest.objectIds}{The \link{DuplicateObjectRequest.objectIds} object or list of objects}

\item{objectIds}{The object being duplicated may contain other objects, for example when}

\item{objectId}{The ID of the object to duplicate}
}
\value{
DuplicateObjectRequest object
}
\description{
DuplicateObjectRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Duplicates a slide or page element.When duplicating a slide, the duplicate slide will be created immediatelyfollowing the specified slide. When duplicating a page element, the duplicatewill be placed on the same page at the same position as the original.
}
\seealso{
Other DuplicateObjectRequest functions: \code{\link{DuplicateObjectRequest.objectIds}}
}
