% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{DuplicateObjectResponse}
\alias{DuplicateObjectResponse}
\title{DuplicateObjectResponse Object}
\usage{
DuplicateObjectResponse(objectId = NULL)
}
\arguments{
\item{objectId}{The ID of the new duplicate object}
}
\value{
DuplicateObjectResponse object
}
\description{
DuplicateObjectResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The response of duplicating an object.
}
