% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Group}
\alias{Group}
\title{Group Object}
\usage{
Group(children = NULL)
}
\arguments{
\item{children}{The collection of elements in the group}
}
\value{
Group object
}
\description{
Group Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representing ajoined collection of PageElements.
}
