% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Image}
\alias{Image}
\title{Image Object}
\usage{
Image(imageProperties = NULL, contentUrl = NULL)
}
\arguments{
\item{imageProperties}{The properties of the image}

\item{contentUrl}{An URL to an image with a default lifetime of 30 minutes}
}
\value{
Image object
}
\description{
Image Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representing animage.
}
