% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ImageProperties}
\alias{ImageProperties}
\title{ImageProperties Object}
\usage{
ImageProperties(shadow = NULL, contrast = NULL, link = NULL,
  recolor = NULL, cropProperties = NULL, outline = NULL,
  brightness = NULL, transparency = NULL)
}
\arguments{
\item{shadow}{The shadow of the image}

\item{contrast}{The contrast effect of the image}

\item{link}{The hyperlink destination of the image}

\item{recolor}{The recolor effect of the image}

\item{cropProperties}{The crop properties of the image}

\item{outline}{The outline of the image}

\item{brightness}{The brightness effect of the image}

\item{transparency}{The transparency effect of the image}
}
\value{
ImageProperties object
}
\description{
ImageProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of the Image.
}
