% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{InsertTableRowsRequest}
\alias{InsertTableRowsRequest}
\title{InsertTableRowsRequest Object}
\usage{
InsertTableRowsRequest(insertBelow = NULL, number = NULL,
  cellLocation = NULL, tableObjectId = NULL)
}
\arguments{
\item{insertBelow}{Whether to insert new rows below the reference cell location}

\item{number}{The number of rows to be inserted}

\item{cellLocation}{The reference table cell location from which rows will be inserted}

\item{tableObjectId}{The table to insert rows into}
}
\value{
InsertTableRowsRequest object
}
\description{
InsertTableRowsRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Inserts rows into a table.
}
