% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{InsertTextRequest}
\alias{InsertTextRequest}
\title{InsertTextRequest Object}
\usage{
InsertTextRequest(objectId = NULL, text = NULL, insertionIndex = NULL,
  cellLocation = NULL)
}
\arguments{
\item{objectId}{The object ID of the shape or table where the text will be inserted}

\item{text}{The text to be inserted}

\item{insertionIndex}{The index where the text will be inserted, in Unicode code units, based}

\item{cellLocation}{The optional table cell location if the text is to be inserted into a table}
}
\value{
InsertTextRequest object
}
\description{
InsertTextRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Inserts text into a shape or a table cell.
}
