% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{LayoutPlaceholderIdMapping}
\alias{LayoutPlaceholderIdMapping}
\title{LayoutPlaceholderIdMapping Object}
\usage{
LayoutPlaceholderIdMapping(layoutPlaceholder = NULL,
  layoutPlaceholderObjectId = NULL, objectId = NULL)
}
\arguments{
\item{layoutPlaceholder}{The placeholder on a layout that will be applied to a slide}

\item{layoutPlaceholderObjectId}{The object ID of the placeholder on a layout that will be applied}

\item{objectId}{A user-supplied object ID for the placeholder identified above that to be}
}
\value{
LayoutPlaceholderIdMapping object
}
\description{
LayoutPlaceholderIdMapping Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The user-specified ID mapping for a placeholder that will be created on aslide from a specified layout.
}
