% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{LayoutProperties}
\alias{LayoutProperties}
\title{LayoutProperties Object}
\usage{
LayoutProperties(masterObjectId = NULL, name = NULL, displayName = NULL)
}
\arguments{
\item{masterObjectId}{The object ID of the master that this layout is based on}

\item{name}{The name of the layout}

\item{displayName}{The human readable name of the layout in the presentation's locale}
}
\value{
LayoutProperties object
}
\description{
LayoutProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of Page are onlyrelevant for pages with page_type LAYOUT.
}
