% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{LayoutReference}
\alias{LayoutReference}
\title{LayoutReference Object}
\usage{
LayoutReference(predefinedLayout = NULL, layoutId = NULL)
}
\arguments{
\item{predefinedLayout}{Predefined layout}

\item{layoutId}{Layout ID: the object ID of one of the layouts in the presentation}
}
\value{
LayoutReference object
}
\description{
LayoutReference Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Slide layout reference. This may reference either:- A predefined layout- One of the layouts in the presentation.
}
