% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Line}
\alias{Line}
\title{Line Object}
\usage{
Line(lineType = NULL, lineProperties = NULL)
}
\arguments{
\item{lineType}{The type of the line}

\item{lineProperties}{The properties of the line}
}
\value{
Line object
}
\description{
Line Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representing aline, curved connector, or bent connector.
}
