% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{LineFill}
\alias{LineFill}
\title{LineFill Object}
\usage{
LineFill(solidFill = NULL)
}
\arguments{
\item{solidFill}{Solid color fill}
}
\value{
LineFill object
}
\description{
LineFill Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The fill of the line.
}
