% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{List.nestingLevel}
\alias{List.nestingLevel}
\title{List.nestingLevel Object}
\usage{
List.nestingLevel()
}
\value{
List.nestingLevel object
}
\description{
List.nestingLevel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A map of nesting levels to the properties of bullets at the associatedlevel. A list has at most nine levels of nesting, so the possible valuesfor the keys of this map are 0 through 8, inclusive.
}
\seealso{
Other List functions: \code{\link{List}}
}
