% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{NestingLevel}
\alias{NestingLevel}
\title{NestingLevel Object}
\usage{
NestingLevel(bulletStyle = NULL)
}
\arguments{
\item{bulletStyle}{The style of a bullet at this level of nesting}
}
\value{
NestingLevel object
}
\description{
NestingLevel Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Contains properties describing the look and feel of a list bullet at a givenlevel of nesting.
}
