% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{NotesProperties}
\alias{NotesProperties}
\title{NotesProperties Object}
\usage{
NotesProperties(speakerNotesObjectId = NULL)
}
\arguments{
\item{speakerNotesObjectId}{The object ID of the shape on this notes page that contains the speaker}
}
\value{
NotesProperties object
}
\description{
NotesProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of Page that are onlyrelevant for pages with page_type NOTES.
}
