% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{OpaqueColor}
\alias{OpaqueColor}
\title{OpaqueColor Object}
\usage{
OpaqueColor(rgbColor = NULL, themeColor = NULL)
}
\arguments{
\item{rgbColor}{An opaque RGB color}

\item{themeColor}{An opaque theme color}
}
\value{
OpaqueColor object
}
\description{
OpaqueColor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A themeable solid color value.
}
