% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{OptionalColor}
\alias{OptionalColor}
\title{OptionalColor Object}
\usage{
OptionalColor(opaqueColor = NULL)
}
\arguments{
\item{opaqueColor}{If set, this will be used as an opaque color}
}
\value{
OptionalColor object
}
\description{
OptionalColor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A color that can either be fully opaque or fully transparent.
}
