% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Page}
\alias{Page}
\title{Page Object}
\usage{
Page(objectId = NULL, layoutProperties = NULL, notesProperties = NULL,
  pageType = NULL, pageElements = NULL, slideProperties = NULL,
  pageProperties = NULL)
}
\arguments{
\item{objectId}{The object ID for this page}

\item{layoutProperties}{Layout specific properties}

\item{notesProperties}{Notes specific properties}

\item{pageType}{The type of the page}

\item{pageElements}{The page elements rendered on the page}

\item{slideProperties}{Slide specific properties}

\item{pageProperties}{The properties of the page}
}
\value{
Page object
}
\description{
Page Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A page in a presentation.
}
