% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{PageBackgroundFill}
\alias{PageBackgroundFill}
\title{PageBackgroundFill Object}
\usage{
PageBackgroundFill(propertyState = NULL, stretchedPictureFill = NULL,
  solidFill = NULL)
}
\arguments{
\item{propertyState}{The background fill property state}

\item{stretchedPictureFill}{Stretched picture fill}

\item{solidFill}{Solid color fill}
}
\value{
PageBackgroundFill object
}
\description{
PageBackgroundFill Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The page background fill.
}
