% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{PageElementProperties}
\alias{PageElementProperties}
\title{PageElementProperties Object}
\usage{
PageElementProperties(transform = NULL, pageObjectId = NULL, size = NULL)
}
\arguments{
\item{transform}{The transform for the element}

\item{pageObjectId}{The object ID of the page where the element is located}

\item{size}{The size of the element}
}
\value{
PageElementProperties object
}
\description{
PageElementProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Common properties for a page element.Note: When you initially create aPageElement, the API may modifythe values of both `size` and `transform`, but thevisual size will be unchanged.
}
