% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{PageProperties}
\alias{PageProperties}
\title{PageProperties Object}
\usage{
PageProperties(pageBackgroundFill = NULL, colorScheme = NULL)
}
\arguments{
\item{pageBackgroundFill}{The background fill of the page}

\item{colorScheme}{The color scheme of the page}
}
\value{
PageProperties object
}
\description{
PageProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of the Page.The page will inherit properties from the parent page. Depending on the pagetype the hierarchy is defined in eitherSlideProperties orLayoutProperties.
}
