% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ParagraphMarker}
\alias{ParagraphMarker}
\title{ParagraphMarker Object}
\usage{
ParagraphMarker(style = NULL, bullet = NULL)
}
\arguments{
\item{style}{The paragraph's style}

\item{bullet}{The bullet for this paragraph}
}
\value{
ParagraphMarker object
}
\description{
ParagraphMarker Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TextElement kind that represents the beginning of a new paragraph.
}
