% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ParagraphStyle}
\alias{ParagraphStyle}
\title{ParagraphStyle Object}
\usage{
ParagraphStyle(spacingMode = NULL, indentEnd = NULL, indentStart = NULL,
  spaceAbove = NULL, alignment = NULL, lineSpacing = NULL,
  indentFirstLine = NULL, spaceBelow = NULL, direction = NULL)
}
\arguments{
\item{spacingMode}{The spacing mode for the paragraph}

\item{indentEnd}{The amount indentation for the paragraph on the side that corresponds to}

\item{indentStart}{The amount indentation for the paragraph on the side that corresponds to}

\item{spaceAbove}{The amount of extra space above the paragraph}

\item{alignment}{The text alignment for this paragraph}

\item{lineSpacing}{The amount of space between lines, as a percentage of normal, where normal}

\item{indentFirstLine}{The amount of indentation for the start of the first line of the paragraph}

\item{spaceBelow}{The amount of extra space above the paragraph}

\item{direction}{The text direction of this paragraph}
}
\value{
ParagraphStyle object
}
\description{
ParagraphStyle Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Styles that apply to a whole paragraph.If this text is contained in a shape with a parent placeholder, then these paragraph styles may beinherited from the parent. Which paragraph styles are inherited depend on thenesting level of lists:* A paragraph not in a list will inherit its paragraph style from the  paragraph at the 0 nesting level of the list inside the parent placeholder.* A paragraph in a list will inherit its paragraph style from the paragraph  at its corresponding nesting level of the list inside the parent  placeholder.Inherited paragraph styles are represented as unset fields in this message.
}
