% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Presentation}
\alias{Presentation}
\title{Presentation Object}
\usage{
Presentation(presentationId = NULL, slides = NULL, notesMaster = NULL,
  layouts = NULL, title = NULL, masters = NULL, locale = NULL,
  pageSize = NULL)
}
\arguments{
\item{presentationId}{The ID of the presentation}

\item{slides}{The slides in the presentation}

\item{notesMaster}{The notes master in the presentation}

\item{layouts}{The layouts in the presentation}

\item{title}{The title of the presentation}

\item{masters}{The slide masters in the presentation}

\item{locale}{The locale of the presentation, as an IETF BCP 47 language tag}

\item{pageSize}{The size of pages in the presentation}
}
\value{
Presentation object
}
\description{
Presentation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A Google Slides presentation.
}
\seealso{
Other Presentation functions: \code{\link{presentations.create}}
}
