% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_functions.R
\name{presentations.batchUpdate}
\alias{presentations.batchUpdate}
\title{Applies one or more updates to the presentation.Each request is validated beforebeing applied. If any request is not valid, then the entire request willfail and nothing will be applied.Some requests have replies togive you some information about how they are applied. Other requests donot need to return information; these each return an empty reply.The order of replies matches that of the requests.For example, suppose you call batchUpdate with four updates, and only thethird one returns information. The response would have two empty replies:the reply to the third request, and another empty reply, in that order.Because other users may be editing the presentation, the presentationmight not exactly reflect your changes: your changes maybe altered with respect to collaborator changes. If there are nocollaborators, the presentation should reflect your changes. In any case,the updates in your request are guaranteed to be applied togetheratomically.}
\usage{
presentations.batchUpdate(BatchUpdatePresentationRequest, presentationId)
}
\arguments{
\item{BatchUpdatePresentationRequest}{The \link{BatchUpdatePresentationRequest} object to pass to this method}

\item{presentationId}{The presentation to apply the updates to}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.readonly
\item https://www.googleapis.com/auth/presentations
\item https://www.googleapis.com/auth/spreadsheets
\item https://www.googleapis.com/auth/spreadsheets.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.readonly, https://www.googleapis.com/auth/presentations, https://www.googleapis.com/auth/spreadsheets, https://www.googleapis.com/auth/spreadsheets.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/slides/}{Google Documentation}

Other BatchUpdatePresentationRequest functions: \code{\link{BatchUpdatePresentationRequest}}
}
