% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_functions.R
\name{presentations.get}
\alias{presentations.get}
\title{Gets the latest version of the specified presentation.}
\usage{
presentations.get(presentationId)
}
\arguments{
\item{presentationId}{The ID of the presentation to retrieve}
}
\description{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_skeleton}}
}
\details{
Authentication scopes used by this function are:
\itemize{
  \item https://www.googleapis.com/auth/drive
\item https://www.googleapis.com/auth/drive.readonly
\item https://www.googleapis.com/auth/presentations
\item https://www.googleapis.com/auth/presentations.readonly
}

Set \code{options(googleAuthR.scopes.selected = c(https://www.googleapis.com/auth/drive, https://www.googleapis.com/auth/drive.readonly, https://www.googleapis.com/auth/presentations, https://www.googleapis.com/auth/presentations.readonly)}
Then run \code{googleAuthR::gar_auth()} to authenticate.
See \code{\link[googleAuthR]{gar_auth}} for details.
}
\seealso{
\href{https://developers.google.com/slides/}{Google Documentation}
}
