% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Recolor}
\alias{Recolor}
\title{Recolor Object}
\usage{
Recolor(recolorStops = NULL, name = NULL)
}
\arguments{
\item{recolorStops}{The recolor effect is represented by a gradient, which is a list of color}

\item{name}{The name of the recolor effect}
}
\value{
Recolor object
}
\description{
Recolor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A recolor effect applied on an image.
}
