% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ReplaceAllShapesWithImageRequest}
\alias{ReplaceAllShapesWithImageRequest}
\title{ReplaceAllShapesWithImageRequest Object}
\usage{
ReplaceAllShapesWithImageRequest(imageUrl = NULL, replaceMethod = NULL,
  containsText = NULL)
}
\arguments{
\item{imageUrl}{The image URL}

\item{replaceMethod}{The replace method}

\item{containsText}{If set, this request will replace all of the shapes that contain the}
}
\value{
ReplaceAllShapesWithImageRequest object
}
\description{
ReplaceAllShapesWithImageRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Replaces all shapes that match the given criteria with the provided image.
}
