% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ReplaceAllShapesWithImageResponse}
\alias{ReplaceAllShapesWithImageResponse}
\title{ReplaceAllShapesWithImageResponse Object}
\usage{
ReplaceAllShapesWithImageResponse(occurrencesChanged = NULL)
}
\arguments{
\item{occurrencesChanged}{The number of shapes replaced with images}
}
\value{
ReplaceAllShapesWithImageResponse object
}
\description{
ReplaceAllShapesWithImageResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of replacing shapes with an image.
}
