% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ReplaceAllShapesWithSheetsChartRequest}
\alias{ReplaceAllShapesWithSheetsChartRequest}
\title{ReplaceAllShapesWithSheetsChartRequest Object}
\usage{
ReplaceAllShapesWithSheetsChartRequest(spreadsheetId = NULL,
  linkingMode = NULL, containsText = NULL, chartId = NULL)
}
\arguments{
\item{spreadsheetId}{The ID of the Google Sheets spreadsheet that contains the chart}

\item{linkingMode}{The mode with which the chart is linked to the source spreadsheet}

\item{containsText}{The criteria that the shapes must match in order to be replaced}

\item{chartId}{The ID of the specific chart in the Google Sheets spreadsheet}
}
\value{
ReplaceAllShapesWithSheetsChartRequest object
}
\description{
ReplaceAllShapesWithSheetsChartRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Replaces all shapes that match the given criteria with the provided GoogleSheets chart. The chart will be scaled and centered to fit within the boundsof the original shape.NOTE: Replacing shapes with a chart requires at least one of thespreadsheets.readonly, spreadsheets, drive.readonly, or drive OAuth scopes.
}
