% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ReplaceAllShapesWithSheetsChartResponse}
\alias{ReplaceAllShapesWithSheetsChartResponse}
\title{ReplaceAllShapesWithSheetsChartResponse Object}
\usage{
ReplaceAllShapesWithSheetsChartResponse(occurrencesChanged = NULL)
}
\arguments{
\item{occurrencesChanged}{The number of shapes replaced with charts}
}
\value{
ReplaceAllShapesWithSheetsChartResponse object
}
\description{
ReplaceAllShapesWithSheetsChartResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of replacing shapes with a Google Sheets chart.
}
