% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ReplaceAllTextRequest}
\alias{ReplaceAllTextRequest}
\title{ReplaceAllTextRequest Object}
\usage{
ReplaceAllTextRequest(replaceText = NULL, containsText = NULL)
}
\arguments{
\item{replaceText}{The text that will replace the matched text}

\item{containsText}{Finds text in a shape matching this substring}
}
\value{
ReplaceAllTextRequest object
}
\description{
ReplaceAllTextRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Replaces all instances of text matching a criteria with replace text.
}
