% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ReplaceAllTextResponse}
\alias{ReplaceAllTextResponse}
\title{ReplaceAllTextResponse Object}
\usage{
ReplaceAllTextResponse(occurrencesChanged = NULL)
}
\arguments{
\item{occurrencesChanged}{The number of occurrences changed by replacing all text}
}
\value{
ReplaceAllTextResponse object
}
\description{
ReplaceAllTextResponse Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The result of replacing text.
}
