% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Request}
\alias{Request}
\title{Request Object}
\usage{
Request(replaceAllShapesWithImage = NULL, replaceAllText = NULL,
  updateImageProperties = NULL, insertTableRows = NULL,
  createSlide = NULL, updateLineProperties = NULL,
  updateSlidesPosition = NULL, deleteTableRow = NULL,
  updateShapeProperties = NULL, insertText = NULL, deleteText = NULL,
  updatePageProperties = NULL, createShape = NULL,
  deleteParagraphBullets = NULL, insertTableColumns = NULL,
  refreshSheetsChart = NULL, createTable = NULL,
  updateTableCellProperties = NULL, deleteObject = NULL,
  updateParagraphStyle = NULL, duplicateObject = NULL,
  deleteTableColumn = NULL, createLine = NULL,
  updateVideoProperties = NULL, createImage = NULL,
  createParagraphBullets = NULL, createVideo = NULL,
  replaceAllShapesWithSheetsChart = NULL, createSheetsChart = NULL,
  updatePageElementTransform = NULL, updateTextStyle = NULL)
}
\arguments{
\item{replaceAllShapesWithImage}{Replaces all shapes matching some criteria with an image}

\item{replaceAllText}{Replaces all instances of specified text}

\item{updateImageProperties}{Updates the properties of an Image}

\item{insertTableRows}{Inserts rows into a table}

\item{createSlide}{Creates a new slide}

\item{updateLineProperties}{Updates the properties of a Line}

\item{updateSlidesPosition}{Updates the position of a set of slides in the presentation}

\item{deleteTableRow}{Deletes a row from a table}

\item{updateShapeProperties}{Updates the properties of a Shape}

\item{insertText}{Inserts text into a shape or table cell}

\item{deleteText}{Deletes text from a shape or a table cell}

\item{updatePageProperties}{Updates the properties of a Page}

\item{createShape}{Creates a new shape}

\item{deleteParagraphBullets}{Deletes bullets from paragraphs}

\item{insertTableColumns}{Inserts columns into a table}

\item{refreshSheetsChart}{Refreshes a Google Sheets chart}

\item{createTable}{Creates a new table}

\item{updateTableCellProperties}{Updates the properties of a TableCell}

\item{deleteObject}{Deletes a page or page element from the presentation}

\item{updateParagraphStyle}{Updates the styling of paragraphs within a Shape or Table}

\item{duplicateObject}{Duplicates a slide or page element}

\item{deleteTableColumn}{Deletes a column from a table}

\item{createLine}{Creates a line}

\item{updateVideoProperties}{Updates the properties of a Video}

\item{createImage}{Creates an image}

\item{createParagraphBullets}{Creates bullets for paragraphs}

\item{createVideo}{Creates a video}

\item{replaceAllShapesWithSheetsChart}{Replaces all shapes matching some criteria with a Google Sheets chart}

\item{createSheetsChart}{Creates an embedded Google Sheets chart}

\item{updatePageElementTransform}{Updates the transform of a page element}

\item{updateTextStyle}{Updates the styling of text within a Shape or Table}
}
\value{
Request object
}
\description{
Request Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single kind of update to apply to a presentation.
}
