% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Response}
\alias{Response}
\title{Response Object}
\usage{
Response(createVideo = NULL, createSheetsChart = NULL,
  replaceAllShapesWithSheetsChart = NULL, replaceAllShapesWithImage = NULL,
  createTable = NULL, replaceAllText = NULL, createSlide = NULL,
  duplicateObject = NULL, createShape = NULL, createLine = NULL,
  createImage = NULL)
}
\arguments{
\item{createVideo}{The result of creating a video}

\item{createSheetsChart}{The result of creating a Google Sheets chart}

\item{replaceAllShapesWithSheetsChart}{The result of replacing all shapes matching some criteria with a Google}

\item{replaceAllShapesWithImage}{The result of replacing all shapes matching some criteria with an}

\item{createTable}{The result of creating a table}

\item{replaceAllText}{The result of replacing text}

\item{createSlide}{The result of creating a slide}

\item{duplicateObject}{The result of duplicating an object}

\item{createShape}{The result of creating a shape}

\item{createLine}{The result of creating a line}

\item{createImage}{The result of creating an image}
}
\value{
Response object
}
\description{
Response Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A single response from an update.
}
