% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{RgbColor}
\alias{RgbColor}
\title{RgbColor Object}
\usage{
RgbColor(green = NULL, blue = NULL, red = NULL)
}
\arguments{
\item{green}{The green component of the color, from 0}

\item{blue}{The blue component of the color, from 0}

\item{red}{The red component of the color, from 0}
}
\value{
RgbColor object
}
\description{
RgbColor Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
An RGB color.
}
