% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Shadow}
\alias{Shadow}
\title{Shadow Object}
\usage{
Shadow(alpha = NULL, color = NULL, rotateWithShape = NULL,
  propertyState = NULL, blurRadius = NULL, type = NULL,
  transform = NULL, alignment = NULL)
}
\arguments{
\item{alpha}{The alpha of the shadow's color, from 0}

\item{color}{The shadow color value}

\item{rotateWithShape}{Whether the shadow should rotate with the shape}

\item{propertyState}{The shadow property state}

\item{blurRadius}{The radius of the shadow blur}

\item{type}{The type of the shadow}

\item{transform}{Transform that encodes the translate, scale, and skew of the shadow,}

\item{alignment}{The alignment point of the shadow, that sets the origin for translate,}
}
\value{
Shadow object
}
\description{
Shadow Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The shadow properties of a page element.If these fields are unset, they may be inherited from a parent placeholderif it exists. If there is no parent, the fields will default to the valueused for new page elements created in the Slides editor, which may depend onthe page element kind.
}
