% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Shape}
\alias{Shape}
\title{Shape Object}
\usage{
Shape(shapeType = NULL, text = NULL, placeholder = NULL,
  shapeProperties = NULL)
}
\arguments{
\item{shapeType}{The type of the shape}

\item{text}{The text content of the shape}

\item{placeholder}{Placeholders are shapes that are inherit from corresponding placeholders on}

\item{shapeProperties}{The properties of the shape}
}
\value{
Shape object
}
\description{
Shape Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representing ageneric shape that does not have a more specific classification.
}
