% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ShapeBackgroundFill}
\alias{ShapeBackgroundFill}
\title{ShapeBackgroundFill Object}
\usage{
ShapeBackgroundFill(propertyState = NULL, solidFill = NULL)
}
\arguments{
\item{propertyState}{The background fill property state}

\item{solidFill}{Solid color fill}
}
\value{
ShapeBackgroundFill object
}
\description{
ShapeBackgroundFill Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The shape background fill.
}
