% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ShapeProperties}
\alias{ShapeProperties}
\title{ShapeProperties Object}
\usage{
ShapeProperties(outline = NULL, shapeBackgroundFill = NULL, shadow = NULL,
  link = NULL)
}
\arguments{
\item{outline}{The outline of the shape}

\item{shapeBackgroundFill}{The background fill of the shape}

\item{shadow}{The shadow properties of the shape}

\item{link}{The hyperlink destination of the shape}
}
\value{
ShapeProperties object
}
\description{
ShapeProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of a Shape.If the shape is a placeholder shape as determined by theplaceholder field, then theseproperties may be inherited from a parent placeholder shape.Determining the rendered value of the property depends on the correspondingproperty_state field value.
}
