% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{SheetsChart}
\alias{SheetsChart}
\title{SheetsChart Object}
\usage{
SheetsChart(sheetsChartProperties = NULL, contentUrl = NULL,
  spreadsheetId = NULL, chartId = NULL)
}
\arguments{
\item{sheetsChartProperties}{The properties of the Sheets chart}

\item{contentUrl}{The URL of an image of the embedded chart, with a default lifetime of 30}

\item{spreadsheetId}{The ID of the Google Sheets spreadsheet that contains the source chart}

\item{chartId}{The ID of the specific chart in the Google Sheets spreadsheet that is}
}
\value{
SheetsChart object
}
\description{
SheetsChart Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representinga linked chart embedded from Google Sheets.
}
