% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{SheetsChartProperties}
\alias{SheetsChartProperties}
\title{SheetsChartProperties Object}
\usage{
SheetsChartProperties(chartImageProperties = NULL)
}
\arguments{
\item{chartImageProperties}{The properties of the embedded chart image}
}
\value{
SheetsChartProperties object
}
\description{
SheetsChartProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of the SheetsChart.
}
