% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{SlideProperties}
\alias{SlideProperties}
\title{SlideProperties Object}
\usage{
SlideProperties(notesPage = NULL, layoutObjectId = NULL,
  masterObjectId = NULL)
}
\arguments{
\item{notesPage}{The notes page that this slide is associated with}

\item{layoutObjectId}{The object ID of the layout that this slide is based on}

\item{masterObjectId}{The object ID of the master that this slide is based on}
}
\value{
SlideProperties object
}
\description{
SlideProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of Page that are onlyrelevant for pages with page_type SLIDE.
}
