% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_functions.R
\docType{package}
\name{slides_googleAuthR}
\alias{slides_googleAuthR}
\alias{slides_googleAuthR-package}
\title{Google Slides API
An API for creating and editing Google Slides presentations.}
\description{
Auto-generated code by googleAuthR::gar_create_api_skeleton
 at 2017-03-05 20:14:17
filename: /Users/mark/dev/R/autoGoogleAPI/googleslidesv1.auto/R/slides_functions.R
api_json: api_json
}
\details{
Authentication scopes used are:
\itemize{
  \item https://www.googleapis.com/auth/spreadsheets
\item https://www.googleapis.com/auth/presentations
\item https://www.googleapis.com/auth/presentations.readonly
\item https://www.googleapis.com/auth/drive.readonly
\item https://www.googleapis.com/auth/spreadsheets.readonly
\item https://www.googleapis.com/auth/drive
}
}
