% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{SolidFill}
\alias{SolidFill}
\title{SolidFill Object}
\usage{
SolidFill(alpha = NULL, color = NULL)
}
\arguments{
\item{alpha}{The fraction of this `color` that should be applied to the pixel}

\item{color}{The color value of the solid fill}
}
\value{
SolidFill object
}
\description{
SolidFill Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A solid color fill. The page or page element is filled entirely with thespecified color value.If any field is unset, its value may be inherited from a parent placeholderif it exists.
}
