% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{StretchedPictureFill}
\alias{StretchedPictureFill}
\title{StretchedPictureFill Object}
\usage{
StretchedPictureFill(contentUrl = NULL, size = NULL)
}
\arguments{
\item{contentUrl}{Reading the content_url:}

\item{size}{The original size of the picture fill}
}
\value{
StretchedPictureFill object
}
\description{
StretchedPictureFill Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The stretched picture fill. The page or page element is filled entirely withthe specified picture. The picture is stretched to fit its container.
}
