% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{SubstringMatchCriteria}
\alias{SubstringMatchCriteria}
\title{SubstringMatchCriteria Object}
\usage{
SubstringMatchCriteria(text = NULL, matchCase = NULL)
}
\arguments{
\item{text}{The text to search for in the shape or table}

\item{matchCase}{Indicates whether the search should respect case:}
}
\value{
SubstringMatchCriteria object
}
\description{
SubstringMatchCriteria Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A criteria that matches a specific string of text in a shape or table.
}
