% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Table}
\alias{Table}
\title{Google Slides API Objects 
An API for creating and editing Google Slides presentations.}
\usage{
Table(rows = NULL, tableColumns = NULL, columns = NULL,
  tableRows = NULL)
}
\arguments{
\item{rows}{Number of rows in the table}

\item{tableColumns}{Properties of each column}

\item{columns}{Number of columns in the table}

\item{tableRows}{Properties and contents of each row}
}
\value{
Table object
}
\description{
Auto-generated code by googleAuthR::gar_create_api_objects
 at 2017-03-05 20:14:17
filename: /Users/mark/dev/R/autoGoogleAPI/googleslidesv1.auto/R/slides_objects.R
api_json: api_json
}
\details{
Objects for use by the functions created by googleAuthR::gar_create_api_skeleton
Table Object


Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representing atable.
}
