% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{TableCellBackgroundFill}
\alias{TableCellBackgroundFill}
\title{TableCellBackgroundFill Object}
\usage{
TableCellBackgroundFill(solidFill = NULL, propertyState = NULL)
}
\arguments{
\item{solidFill}{Solid color fill}

\item{propertyState}{The background fill property state}
}
\value{
TableCellBackgroundFill object
}
\description{
TableCellBackgroundFill Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The table cell background fill.
}
