% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{TableCellLocation}
\alias{TableCellLocation}
\title{TableCellLocation Object}
\usage{
TableCellLocation(rowIndex = NULL, columnIndex = NULL)
}
\arguments{
\item{rowIndex}{The 0-based row index}

\item{columnIndex}{The 0-based column index}
}
\value{
TableCellLocation object
}
\description{
TableCellLocation Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A location of a single table cell within a table.
}
