% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{TableColumnProperties}
\alias{TableColumnProperties}
\title{TableColumnProperties Object}
\usage{
TableColumnProperties(columnWidth = NULL)
}
\arguments{
\item{columnWidth}{Width of a column}
}
\value{
TableColumnProperties object
}
\description{
TableColumnProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties of each column in a table.
}
