% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{TableRange}
\alias{TableRange}
\title{TableRange Object}
\usage{
TableRange(location = NULL, rowSpan = NULL, columnSpan = NULL)
}
\arguments{
\item{location}{The starting location of the table range}

\item{rowSpan}{The row span of the table range}

\item{columnSpan}{The column span of the table range}
}
\value{
TableRange object
}
\description{
TableRange Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A table range represents a reference to a subset of a table.It's important to note that the cells specified by a table range do notnecessarily form a rectangle. For example, let's say we have a 3 x 3 tablewhere all the cells of the last row are merged together. The table lookslike this:             [             ]A table range with location = (0, 0), row span = 3 and column span = 2specifies the following cells:   x     x   [      x      ]
}
