% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{TableRow}
\alias{TableRow}
\title{TableRow Object}
\usage{
TableRow(rowHeight = NULL, tableCells = NULL)
}
\arguments{
\item{rowHeight}{Height of a row}

\item{tableCells}{Properties and contents of each cell}
}
\value{
TableRow object
}
\description{
TableRow Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Properties and contents of each row in a table.
}
