% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{TextContent.lists}
\alias{TextContent.lists}
\title{TextContent.lists Object}
\usage{
TextContent.lists()
}
\value{
TextContent.lists object
}
\description{
TextContent.lists Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The bulleted lists contained in this text, keyed by list ID.
}
\seealso{
Other TextContent functions: \code{\link{TextContent}}
}
