% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{TextElement}
\alias{TextElement}
\title{TextElement Object}
\usage{
TextElement(autoText = NULL, paragraphMarker = NULL, startIndex = NULL,
  endIndex = NULL, textRun = NULL)
}
\arguments{
\item{autoText}{A TextElement representing a spot in the text that is dynamically}

\item{paragraphMarker}{A marker representing the beginning of a new paragraph}

\item{startIndex}{The zero-based start index of this text element, in Unicode code units}

\item{endIndex}{The zero-based end index of this text element, exclusive, in Unicode code}

\item{textRun}{A TextElement representing a run of text where all of the characters}
}
\value{
TextElement object
}
\description{
TextElement Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TextElement describes the content of a range of indices in the text contentof a Shape or TableCell.
}
