% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{TextRun}
\alias{TextRun}
\title{TextRun Object}
\usage{
TextRun(style = NULL, content = NULL)
}
\arguments{
\item{style}{The styling applied to this run}

\item{content}{The text of this run}
}
\value{
TextRun object
}
\description{
TextRun Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A TextElement kind that represents a run of text that all has the samestyling.
}
