% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{ThemeColorPair}
\alias{ThemeColorPair}
\title{ThemeColorPair Object}
\usage{
ThemeColorPair(color = NULL, type = NULL)
}
\arguments{
\item{color}{The concrete color corresponding to the theme color type above}

\item{type}{The type of the theme color}
}
\value{
ThemeColorPair object
}
\description{
ThemeColorPair Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A pair mapping a theme color type to the concrete color it represents.
}
