% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{UpdateImagePropertiesRequest}
\alias{UpdateImagePropertiesRequest}
\title{UpdateImagePropertiesRequest Object}
\usage{
UpdateImagePropertiesRequest(objectId = NULL, fields = NULL,
  imageProperties = NULL)
}
\arguments{
\item{objectId}{The object ID of the image the updates are applied to}

\item{fields}{The fields that should be updated}

\item{imageProperties}{The image properties to update}
}
\value{
UpdateImagePropertiesRequest object
}
\description{
UpdateImagePropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Update the properties of an Image.
}
