% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{UpdateLinePropertiesRequest}
\alias{UpdateLinePropertiesRequest}
\title{UpdateLinePropertiesRequest Object}
\usage{
UpdateLinePropertiesRequest(objectId = NULL, lineProperties = NULL,
  fields = NULL)
}
\arguments{
\item{objectId}{The object ID of the line the update is applied to}

\item{lineProperties}{The line properties to update}

\item{fields}{The fields that should be updated}
}
\value{
UpdateLinePropertiesRequest object
}
\description{
UpdateLinePropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the properties of a Line.
}
