% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{UpdatePageElementTransformRequest}
\alias{UpdatePageElementTransformRequest}
\title{UpdatePageElementTransformRequest Object}
\usage{
UpdatePageElementTransformRequest(objectId = NULL, transform = NULL,
  applyMode = NULL)
}
\arguments{
\item{objectId}{The object ID of the page element to update}

\item{transform}{The input transform matrix used to update the page element}

\item{applyMode}{The apply mode of the transform update}
}
\value{
UpdatePageElementTransformRequest object
}
\description{
UpdatePageElementTransformRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the transform of a page element.
}
