% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{UpdatePagePropertiesRequest}
\alias{UpdatePagePropertiesRequest}
\title{UpdatePagePropertiesRequest Object}
\usage{
UpdatePagePropertiesRequest(fields = NULL, objectId = NULL,
  pageProperties = NULL)
}
\arguments{
\item{fields}{The fields that should be updated}

\item{objectId}{The object ID of the page the update is applied to}

\item{pageProperties}{The page properties to update}
}
\value{
UpdatePagePropertiesRequest object
}
\description{
UpdatePagePropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the properties of a Page.
}
