% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{UpdateParagraphStyleRequest}
\alias{UpdateParagraphStyleRequest}
\title{UpdateParagraphStyleRequest Object}
\usage{
UpdateParagraphStyleRequest(cellLocation = NULL, style = NULL,
  fields = NULL, objectId = NULL, textRange = NULL)
}
\arguments{
\item{cellLocation}{The location of the cell in the table containing the paragraph(s) to}

\item{style}{The paragraph's style}

\item{fields}{The fields that should be updated}

\item{objectId}{The object ID of the shape or table with the text to be styled}

\item{textRange}{The range of text containing the paragraph(s) to style}
}
\value{
UpdateParagraphStyleRequest object
}
\description{
UpdateParagraphStyleRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the styling for all of the paragraphs within a Shape or Table thatoverlap with the given text index range.
}
