% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{UpdateShapePropertiesRequest}
\alias{UpdateShapePropertiesRequest}
\title{UpdateShapePropertiesRequest Object}
\usage{
UpdateShapePropertiesRequest(fields = NULL, objectId = NULL,
  shapeProperties = NULL)
}
\arguments{
\item{fields}{The fields that should be updated}

\item{objectId}{The object ID of the shape the updates are applied to}

\item{shapeProperties}{The shape properties to update}
}
\value{
UpdateShapePropertiesRequest object
}
\description{
UpdateShapePropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Update the properties of a Shape.
}
