% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{UpdateSlidesPositionRequest}
\alias{UpdateSlidesPositionRequest}
\title{UpdateSlidesPositionRequest Object}
\usage{
UpdateSlidesPositionRequest(slideObjectIds = NULL, insertionIndex = NULL)
}
\arguments{
\item{slideObjectIds}{The IDs of the slides in the presentation that should be moved}

\item{insertionIndex}{The index where the slides should be inserted, based on the slide}
}
\value{
UpdateSlidesPositionRequest object
}
\description{
UpdateSlidesPositionRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Updates the position of slides in the presentation.
}
