% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{UpdateTableCellPropertiesRequest}
\alias{UpdateTableCellPropertiesRequest}
\title{UpdateTableCellPropertiesRequest Object}
\usage{
UpdateTableCellPropertiesRequest(fields = NULL, objectId = NULL,
  tableRange = NULL, tableCellProperties = NULL)
}
\arguments{
\item{fields}{The fields that should be updated}

\item{objectId}{The object ID of the table}

\item{tableRange}{The table range representing the subset of the table to which the updates}

\item{tableCellProperties}{The table cell properties to update}
}
\value{
UpdateTableCellPropertiesRequest object
}
\description{
UpdateTableCellPropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Update the properties of a TableCell.
}
