% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{UpdateTextStyleRequest}
\alias{UpdateTextStyleRequest}
\title{UpdateTextStyleRequest Object}
\usage{
UpdateTextStyleRequest(fields = NULL, objectId = NULL, textRange = NULL,
  cellLocation = NULL, style = NULL)
}
\arguments{
\item{fields}{The fields that should be updated}

\item{objectId}{The object ID of the shape or table with the text to be styled}

\item{textRange}{The range of text to style}

\item{cellLocation}{The location of the cell in the table containing the text to style}

\item{style}{The style(s) to set on the text}
}
\value{
UpdateTextStyleRequest object
}
\description{
UpdateTextStyleRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Update the styling of text in a Shape orTable.
}
