% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{UpdateVideoPropertiesRequest}
\alias{UpdateVideoPropertiesRequest}
\title{UpdateVideoPropertiesRequest Object}
\usage{
UpdateVideoPropertiesRequest(fields = NULL, objectId = NULL,
  videoProperties = NULL)
}
\arguments{
\item{fields}{The fields that should be updated}

\item{objectId}{The object ID of the video the updates are applied to}

\item{videoProperties}{The video properties to update}
}
\value{
UpdateVideoPropertiesRequest object
}
\description{
UpdateVideoPropertiesRequest Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
Update the properties of a Video.
}
