% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{Video}
\alias{Video}
\title{Video Object}
\usage{
Video(url = NULL, id = NULL, videoProperties = NULL, source = NULL)
}
\arguments{
\item{url}{An URL to a video}

\item{id}{The video source's unique identifier for this video}

\item{videoProperties}{The properties of the video}

\item{source}{The video source}
}
\value{
Video object
}
\description{
Video Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
A PageElement kind representing avideo.
}
