% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slides_objects.R
\name{VideoProperties}
\alias{VideoProperties}
\title{VideoProperties Object}
\usage{
VideoProperties(outline = NULL)
}
\arguments{
\item{outline}{The outline of the video}
}
\value{
VideoProperties object
}
\description{
VideoProperties Object
}
\details{
Autogenerated via \code{\link[googleAuthR]{gar_create_api_objects}}
The properties of the Video.
}
